backend/requirements.txt
alabaster==0.7.16
alembic==1.13.1
aniso8601==9.0.1
anytree==2.12.1
apispec==6.6.0
apispec-webframeworks==1.1.0
attrs==23.2.0
autopage==0.5.2
Babel==2.14.0
beautifulsoup4==4.12.3
blinker==1.7.0
bullet==2.2.0
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cliff==4.6.0
cmd2==2.4.3
cmdstanpy==1.2.1
colorama==0.4.6
configobj==5.0.8
configparser==6.0.1
contourpy==1.2.0
convertdate==2.4.0
cycler==0.12.1
distlib==0.3.8
dnspython==2.0.0
docutils==0.20.1
ephem==4.1.5
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.50.0
gnupg==2.3.1
greenlet==3.0.3
gunicorn==21.2.0
hijri-converter==2.3.1
holidays==0.45
humanfriendly==10.0
idna==3.6
imagesize==1.4.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipwhois==1.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
LunarCalendar==0.0.9
lxml==5.1.0
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
matplotlib==3.8.3
mccabe==0.7.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
passlib==1.7.4
pbr==6.0.0
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
prettytable==3.10.0
property-manager==3.0
prophet==1.1.5
psutil==5.9.8
ptyprocess==0.7.0
py==1.11.0
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
PyMeeus==0.5.12
pyparsing==3.1.2
pyperclip==1.8.2
pyproject-api==1.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-secrets==23.4.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools-git==1.2
setuptools-scm==8.0.4
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.28
SQLAlchemy-Utils==0.41.1
stanio==0.3.0
stevedore==5.2.0
tomli==2.0.1
tox==4.14.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
verboselogs==1.7
virtualenv==20.25.1
wcwidth==0.2.13
Werkzeug==3.0.1
xkcdpass==1.19.8
yamlreader==3.0.4
zipp==3.18.1
plotly
